move transform code into one script
