+++ packages/language-service/lib/documentFeatures/selectionRanges.ts
@@ -1,9 +1,9 @@
 import type { ServiceContext } from '../types';
 import { languageFeatureWorker } from '../utils/featureWorkers';
-import * as transformer from '../transformer';
 import type * as vscode from 'vscode-languageserver-protocol';
 import { isInsideRange, notEmpty } from '../utils/common';
 import { NoneCancellationToken } from '../utils/cancellation';
+import { transformSelectionRanges } from '../utils/transform';
 
 export function register(context: ServiceContext) {
 
@@ -34,7 +34,7 @@ export function register(context: ServiceContext) {
 				if (!map) {
 					return data;
 				}
-				return transformer.asSelectionRanges(
+				return transformSelectionRanges(
 					data,
 					range => map.toSourceRange(range, data => data.selectionRanges ?? true)
 				);
