--- packages/language-service/lib/transformer/index.ts
@@ -1,11 +1,0 @@
-export { transform as asCompletionItem } from './completionItem';
-export { transform as asCompletionList } from './completionList';
-export { transform as asFoldingRanges } from './foldingRanges';
-export { transform as asHover } from './hover';
-export { transform as asLocation } from './locationLike';
-export { transform as asLocations } from './locationsLike';
-export { transform as asSelectionRange } from './selectionRange';
-export { transform as asSelectionRanges } from './selectionRanges';
-export { transform as asTextEdit } from './textEdit';
-export { transform as asDocumentSymbol } from './documentSymbol';
-export { transform as asWorkspaceSymbol } from './workspaceSymbol';
