--- packages/language-service/lib/transformer/workspaceSymbol.ts
@@ -1,15 +1,0 @@
-import type * as vscode from 'vscode-languageserver-protocol';
-
-export function transform(symbol: vscode.WorkspaceSymbol, getOtherLocation: (location: vscode.Location) => vscode.Location | undefined): vscode.WorkspaceSymbol | undefined {
-	if (!('range' in symbol.location)) {
-		return symbol;
-	}
-	const loc = getOtherLocation(symbol.location);
-	if (!loc) {
-		return;
-	}
-	return {
-		...symbol,
-		location: loc,
-	};
-}
